# shared/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PeoplePay360 — HR & Payroll API
  description: |
    Complete REST API for the PeoplePay360 HR & Payroll Platform.
    Covers Auth, Employees, Contracts, Working Schedules, Time Off,
    Attendance, Salary Structures & Rules, Payruns, Payslips, and Dashboard.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development

# ──────────────────────────────────────────────
# SECURITY
# ──────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ────────────────────────────────────────────
  schemas:
...
    EmployeeStatus:
      type: string
      enum: [ACTIVE, ON_LEAVE, TERMINATED]
    ContractStatus:
      type: string
      enum: [DRAFT, ACTIVE, EXPIRED, CANCELLED]
    DayOfWeek:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
...
    AttendanceStatus:
      type: string
      enum: [PRESENT, ABSENT, PARTIAL]
...
    EmployeeCreate:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        department:
          type: string
        jobPosition:
          type: string
        managerId:
          type: string
          format: uuid
        scheduleId:
          type: string
          format: uuid
        bankName:
          type: string
        bankAccount:
          type: string
        bankIFSC:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        address:
          type: string
        emergencyContact:
          type: string
...
    EmployeeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        jobPosition:
          type: string
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        managerId:
          type: string
          nullable: true
        scheduleId:
          type: string
          nullable: true
        bankName:
          type: string
        bankAccount:
          type: string
        bankIFSC:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- Contract ---
    ContractCreate:
      type: object
      required: [employeeId, name, startDate, wage]
      properties:
        employeeId:
          type: string
          format: uuid
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        wage:
          type: number
        department:
          type: string
        jobPosition:
          type: string
        salaryStructureId:
          type: string
          format: uuid
        scheduleId:
          type: string
          format: uuid
        notes:
          type: string
...
    ContractResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        wage:
          type: number
        department:
          type: string
        jobPosition:
          type: string
        status:
          $ref: '#/components/schemas/ContractStatus'
        salaryStructureId:
          type: string
          nullable: true
        scheduleId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- Working Schedule ---
    ScheduleLineInput:
      type: object
      required: [day, startTime, endTime]
      properties:
        day:
          $ref: '#/components/schemas/DayOfWeek'
        startTime:
          type: string
          example: "09:00"
        endTime:
          type: string
          example: "17:00"
        breakMins:
          type: integer
          default: 0
    WorkingScheduleCreate:
      type: object
      required: [name, lines]
      properties:
        name:
          type: string
        type:
          type: string
          default: standard
        lines:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleLineInput'
...
    WorkingScheduleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        totalHours:
          type: number
        isActive:
          type: boolean
        lines:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              day:
                $ref: '#/components/schemas/DayOfWeek'
              startTime:
                type: string
              endTime:
                type: string
              breakMins:
                type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- Time Off Type ---
...
    AttendanceCreate:
      type: object
      required: [employeeId, checkIn]
      properties:
        employeeId:
          type: string
          format: uuid
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
...
    AttendanceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
          nullable: true
        workedHours:
          type: number
          nullable: true
        status:
          $ref: '#/components/schemas/AttendanceStatus'
        isCorrected:
          type: boolean
        originalCheckIn:
          type: string
          format: date-time
          nullable: true
        originalCheckOut:
          type: string
          format: date-time
          nullable: true
        correctedBy:
          type: string
          nullable: true
        correctionNote:
          type: string
          nullable: true

    # --- Salary Structure ---
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
    DepartmentFilter:
      name: department
      in: query
      schema:
        type: string
    StatusFilter:
      name: status
      in: query
      schema:
        type: string

# ──────────────────────────────────────────────
# PATHS
# ──────────────────────────────────────────────
paths:
...
  /employees:
    get:
      tags: [Employees]
      summary: List all employees (paginated, filterable)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/DepartmentFilter'
        - $ref: '#/components/parameters/StatusFilter'
      responses:
        '200':
          description: Paginated list of employees
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EmployeeResponse'
    post:
      tags: [Employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
...
  /contracts:
    get:
      tags: [Contracts]
      summary: List all contracts (filterable by employee, status)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/StatusFilter'
      responses:
        '200':
          description: Paginated list of contracts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ContractResponse'
    post:
      tags: [Contracts]
      summary: Create a new contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractCreate'
      responses:
        '201':
          description: Contract created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
        '409':
          description: Active contract already exists for this employee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /schedules:
    get:
      tags: [Working Schedules]
      summary: List all working schedules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of working schedules
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/WorkingScheduleResponse'
    post:
      tags: [Working Schedules]
      summary: Create a working schedule with lines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkingScheduleCreate'
      responses:
        '201':
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingScheduleResponse'
...
  /attendances:
    get:
      tags: [Attendance]
      summary: List attendance records (filterable)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated attendance list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AttendanceResponse'
    post:
      tags: [Attendance]
      summary: Create an attendance record (check in)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceCreate'
      responses:
        '201':
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
...